project upgrade - upgrade to spring-boot 2.6.5; - upgrade to spring-cloud 2021.0.1; - upgrade to springdoc-openapi 1.6.6; - upgrade to vault docker image 1.10.0; - upgrade to consul docker image 1.11.4; - upgrade to cassandra docker image 4.0.3; - add docker-build.sh and my-functions.sh; - update remove-docker-images.sh; - update READMEs.

+++ remove-docker-images.sh
@@ -1,7 +1,44 @@
 #!/usr/bin/env bash
 
-docker rmi ivanfranchin/movie-service:1.0.0
-docker rmi ivanfranchin/restaurant-service:1.0.0
-docker rmi ivanfranchin/student-service:1.0.0
-docker rmi ivanfranchin/book-service:1.0.0
-docker rmi springboot-vault-examples_cassandra
+source my-functions.sh
+
+check_script_input_parameter $1
+
+if [ "$1" = "spring-vault-approle-mysql" ] ||
+   [ "$1" = "all" ];
+then
+
+  docker rmi ivanfranchin/movie-service:1.0.0
+
+fi
+
+if [ "$1" = "spring-vault-approle-multi-datasources-mysql" ] ||
+   [ "$1" = "all" ];
+then
+
+  docker rmi ivanfranchin/restaurant-service:1.0.0
+
+fi
+
+if [ "$1" = "spring-cloud-vault-approle-mysql" ] ||
+   [ "$1" = "all" ];
+then
+
+  docker rmi ivanfranchin/student-service:1.0.0
+
+fi
+
+if [ "$1" = "spring-cloud-vault-approle-cassandra" ] ||
+   [ "$1" = "all" ];
+then
+
+  docker rmi ivanfranchin/book-service:1.0.0
+
+fi
+
+if [ "$1" = "all" ];
+then
+
+  docker rmi springboot-vault-examples_cassandra
+
+fi
